Add gifs to ReadMe
