+++ README.md
@@ -3,20 +3,22 @@
 Hold up a looking glass to the world, and transform the view! Emblem presents an augmented reality window, displaying geocached two- and three- dimensional models onto users' surroundings.
 
 Web client is available for use at [emblemAR.com](https://www.emblemar.com/).
-Mobile app can be downloaded from the Apple App store at [TODO](http://www.apple.com/itunes/charts/free-apps/).
+Mobile app can be downloaded from the Apple App store at [TODO](http://www.apple.com/itunes/charts/free-apps/), and can be seen in action on [YouTube](https://youtu.be/dP8gMirOFRw).
 
-[TODO: video of mobile app in action](https://github.com/rubensousa/ViewPagerCards)
+<img src="/screencaptures/emblem-raptor.gif" width=300></img><img src="/screencaptures/emblem-statue-of-liberty.gif" width=300></img>
 
 TODO: Update travis-ci build status links (below)
 
-#### [Server and Web Client](https://github.com/Hadashco/emblem-web)
+### [Server and Web Client](https://github.com/Hadashco/emblem-web)
 Upload 3D models from a computer desktop, explore artwork posted around the world, and manage granular user settings from the web client.
+
 [![Build Status](https://api.travis-ci.org/hadashco/emblem-web.svg?branch=master)](https://api.travis-ci.org/hadashco/emblem-web.svg?branch=master)
 [![Stories in Ready](https://badge.waffle.io/Hadashco/emblem-web.png?label=ready&title=Ready)](https://waffle.io/Hadashco/emblem-web)
 
-#### [Mobile App](https://github.com/Hadashco/emblem-mobile)
+### [Mobile App](https://github.com/Hadashco/emblem-mobile)
 In addition to providing a dramatic visual experience, the mobile application offers a well-executed implementation of knitting
 Vuforia's C++ library with Swift using Objective-C.
+
 [![Build Status](https://api.travis-ci.org/hadashco/emblem-mobile.svg?branch=master)](https://api.travis-ci.org/hadashco/emblem-mobile.svg?branch=master)
 [![Stories in Ready](https://badge.waffle.io/Hadashco/emblem-mobile.png?label=ready&title=Ready)](https://waffle.io/Hadashco/emblem-mobile)
